Use generateSequence instead of fromElements for zero and rand vectors
Using fromElements with a List of random or zero elements can cause
timeouts for large parameter values because those values have to be sent
to Flink before the job is run. Instead, generate a sequence of the
desired length and assign a value to each of the elements; this should
solve the timeout issue.